Rename Files
- Circles -> CircleIndicators
- WearSync -> DataProgressItem
- DataProgress -> DataProgressList
- ContributionsMenu -> ContributionsList
